remove s3 support from within hnsqlite
It seems more natural to persist db file outside the hnsqlite layer.
(E.g. will want to support blobstore or other backup mechanism and dont want them all in here.)
Also this sets the stage for hnsqlite not 'owning' the database but rather riding on top of an sqlite file owned by the encompassing system.